refactor tests for asset create hw
